third_party/WebKit/LayoutTests/fast/dom/viewport/viewport-dimensions-under-browser-zoom.html

<!DOCTYPE html>
<style>
  body {
    height: 2000px;
    width: 2000px;
  }
</style>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
  var browserZoomFactor = 1.25;

  test(function() {
    window.scrollTo(100, 100);

    // TODO(ymalik): Remove hook to internals to pinch-zoom here and browser
    // zoom below. This will be required to upstream to w3c repo.
    internals.setPageScaleFactor(2);

    // Offset visual viewport by 10px.
    document.visualViewport.scrollTop = 10;
    document.visualViewport.scrollLeft = 10;

    assert_equals(document.visualViewport.scrollTop, 10);
    assert_equals(document.visualViewport.scrollLeft, 10);
    assert_equals(document.visualViewport.clientWidth, 400);
    assert_equals(document.visualViewport.clientHeight, 300);
    assert_equals(document.visualViewport.pageScale, 2);

    // Apply browser zoom.
    window.internals.setZoomFactor(browserZoomFactor);

    // Visual viewport dimensions should be scaled by the browser zoom factor.
    assert_equals(document.visualViewport.scrollTop, 10 / browserZoomFactor);
    assert_equals(document.visualViewport.scrollLeft, 10 / browserZoomFactor);
    assert_equals(document.visualViewport.clientWidth, 400 / browserZoomFactor);
    assert_equals(document.visualViewport.clientHeight, 300 / browserZoomFactor);
    assert_equals(document.visualViewport.pageScale, 2);

    // Verify setting the visual viewport offset under browser zoom.
    document.visualViewport.scrollTop = 12.5;
    document.visualViewport.scrollLeft = 12.5;
    assert_equals(document.visualViewport.scrollTop, 12.5);
    assert_equals(document.visualViewport.scrollLeft, 12.5);
    assert_equals(document.visualViewport.clientWidth, 400 / browserZoomFactor);
    assert_equals(document.visualViewport.clientHeight, 300 / browserZoomFactor);
    assert_equals(document.visualViewport.pageScale, 2);
  }, 'Verify viewport dimensions under bowser zoom.');
</script>
